refactor(test): org application tests to own module
Extract the application-oriented tests out to a separate module, as the
current one mixes concerns and has become a little unwieldy.

Updates a couple of tests to remove reliance on a route mapping
function when not needed.

Updates test utility import paths to absolute instead of relative.
